requirements.txt
django==3.1.5
django-ckeditor==6.1.0